Bug issue resolved
The number of columns randomly generated was not stable from time to time. 
I hopefully found the solution. 
It is right now appending all the permutation column-wise (so that we have a very long row in the end). 
I did this because of our to_row method which would also be replaced with to_batch method in the nearest future.